gan/triton_server/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
brotlipy=0.7.0=py38h27cfd23_1003
byol-pytorch=0.5.4=pypi_0
ca-certificates=2021.1.19=h06a4308_1
certifi=2020.12.5=py38h578d9bd_1
cffi=1.14.5=py38ha65f79e_0
chardet=4.0.0=py38h06a4308_1003
cloudpickle=1.6.0=py_0
cryptography=3.4.7=py38hd23ed53_0
cudatoolkit=11.1.1=h6406543_8
cycler=0.10.0=py38_0
cytoolz=0.11.0=py38h7b6447c_0
dask-core=2021.2.0=pyhd3eb1b0_0
decorator=4.4.2=pyhd3eb1b0_0
deprecated=1.2.12=pypi_0
dominate=2.6.0=pypi_0
eigen=3.2.8=2
expat=2.2.10=he6710b0_2
fftw=3.3.9=h27cfd23_1
freetype=2.10.4=h0708190_1
future=0.18.2=py38h578d9bd_3
gevent=21.1.2=pypi_0
geventhttpclient=1.4.4=pypi_0
greenlet=1.0.0=pypi_0
hdf5=1.10.6=hb1b8bf9_0
humanize=3.3.0=pypi_0
idna=2.10=pyhd3eb1b0_0
imageio=2.9.0=py_0
ipython=7.21.0=py38h81c977d_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.0=py38h06a4308_1
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9d=h36c2ea0_0
jsonpatch=1.32=pypi_0
jsonpickle=2.0.0=pypi_0
jsonpointer=2.1=pypi_0
kiwisolver=1.3.1=py38h2531618_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h2828fa1_18
libgfortran-ng=7.3.0=hdf63c60_0
libitk=5.1.2=h86d312b_0
libitk-devel=5.1.2=h9c3ff4c_0
liblapack=3.9.0=8_mkl
libpng=1.6.37=h21135ba_2
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=h6de172a_18
libtiff=4.2.0=hdc55705_0
libuv=1.41.0=h7f98852_0
libwebp-base=1.2.0=h7f98852_0
llvm-openmp=11.0.1=h4bd325d_0
lz4-c=1.9.3=h9c3ff4c_0
matplotlib-base=3.3.4=py38h62a2d02_0
mkl=2020.4=h726a3e6_304
mkl-service=2.3.0=py38he904b0f_0
ncurses=6.2=he6710b0_1
networkx=2.5=py_0
nibabel=3.2.1=pypi_0
ninja=1.10.2=h4bd325d_0
numpy=1.20.1=py38h18fd61f_0
nvidia-pyindex=1.0.8=pypi_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1k=h27cfd23_0
packaging=20.9=pypi_0
parso=0.8.1=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.1=py38ha0e1e83_0
pip=21.0.1=py38h06a4308_0
prompt-toolkit=3.0.8=py_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.20=pyh9f0ad1d_2
pygments=2.8.0=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.8=hdb3f193_4
python-dateutil=2.8.1=pyhd3eb1b0_0
python-rapidjson=1.0=pypi_0
python_abi=3.8=1_cp38
pytorch=1.8.0=py3.8_cuda11.1_cudnn8.0.5_0
pywavelets=1.1.1=py38h7b6447c_2
pyyaml=5.4.1=py38h27cfd23_1
pyzmq=20.0.0=py38h2531618_1
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
scikit-image=0.17.2=py38hdf5156a_0
scikit-learn=0.24.1=py38ha9443f7_0
scipy=1.6.1=py38h91f5cce_0
setuptools=52.0.0=py38h06a4308_0
simpleitk=2.0.2=py38hb8ccc7d_1
six=1.15.0=pyh9f0ad1d_0
sqlite=3.33.0=h62c20be_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.10.1=py38hdd07704_2
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.8.0=py38
torchfile=0.1.0=py_0
torchvision=0.9.0=pypi_0
tornado=6.1=py38h27cfd23_0
tqdm=4.56.0=pyhd3eb1b0_0
traitlets=5.0.5=pyhd3eb1b0_0
tritonclient=2.8.0=pypi_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.4=pyhd3eb1b0_0
visdom=0.1.8.9=0
wcwidth=0.2.5=py_0
websocket-client=0.58.0=py38h06a4308_4
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zlib=1.2.11=0
zope-event=4.5.0=pypi_0
zope-interface=5.4.0=pypi_0
zstd=1.4.9=ha95c52a_0
